Refactor MmrCard and MmrTag
The MmrCard and MmrTag components have been replaced by SeasonRankCard
and SeasonRankTag respectively. These components will be used in the
future to display season ranks instead of mmr.
Additionally, the PlayerScreen has been reorganized to fit its child
components a little more neatly. Also, some of the logic for processing
player mmr data has been moved to up to the player screen so that the
result can be passed to the SeasonRankCard.